% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{celldensities}
\alias{celldensities}
\title{Title}
\usage{
celldensities(DCres, density_info)
}
\arguments{
\item{DCres}{TODO}

\item{density_info}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
