% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{check_signature}
\alias{check_signature}
\title{TODO: even if unexported, it would not hurt to have the functions a little more
documented}
\usage{
check_signature(signature_matrix, mix_mat)
}
\arguments{
\item{signature_matrix}{TODO}

\item{mix_mat}{TODO}
}
\description{
TODO: even if unexported, it would not hurt to have the functions a little more
documented
}
