% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr-pkg.R
\docType{data}
\name{dataset_racle}
\alias{dataset_racle}
\title{An exemplary dataset with samples from four patients with metastatic melanoma}
\description{
An exemplary dataset with samples from four patients with metastatic melanoma
}
\details{
\code{quantiseqr} ships with an example dataset with samples from four
patients with metastatic melanoma. The dataset \code{quantiseqr::dataset_racle} contains
\itemize{
\item a gene expression matrix (\code{dataset_racle$expr_mat}) generated using bulk RNA-seq
\item 'gold standard' estimates of immune cell contents profiled with FACS
(\code{dataset_racle$ref}).
}
}
\references{
Racle et al, 2017 - https://doi.org/10.7554/eLife.26476.049
}
