% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{DClsei}
\alias{DClsei}
\title{Title}
\usage{
DClsei(b, A, G, H, scaling)
}
\arguments{
\item{b}{TODO}

\item{A}{TODO}

\item{G}{TODO}

\item{H}{TODO}

\item{scaling}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
