% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{fixMixture}
\alias{fixMixture}
\title{TODO: something similar, but for the SummarizedExperiment class?
TODO: even if unexported, it would not hurt to have the functions a little more
documented}
\usage{
fixMixture(mix.mat, arrays = FALSE)
}
\description{
TODO: something similar, but for the SummarizedExperiment class?
TODO: even if unexported, it would not hurt to have the functions a little more
documented
}
