% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{get_densities}
\alias{get_densities}
\title{Title}
\usage{
get_densities(DCres, density_info)
}
\arguments{
\item{DCres}{Data.frame of deconvoluted cell fractions computed with the "run_quantiseq" function, with sample identifiers as row names.}

\item{density_info}{Named numeric vector of total cell densities per sample. The vector names should match the sample identifiers specified in DCres.
derived from imaging data}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
