% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{makeQN}
\alias{makeQN}
\title{Perform quantile normalization of expression data}
\usage{
makeQN(mix.mat)
}
\arguments{
\item{mix.mat}{Matrix or data.frame with microarray
gene expression values for all samples to be deconvoluted,
with gene symbols as row names and sample IDs as column names.
Expression levels should be on non-log scale.}
}
\value{
The input matrix transformed with quantile normalization.
}
\description{
Perform quantile normalization of expression data
}
\examples{

data(dataset_racle)
# mixture.quantile <- quantiseqr:::makeQN(dataset_racle$expr_mat)
}
