% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{mapGenes}
\alias{mapGenes}
\title{mapGenes}
\usage{
mapGenes(mydata)
}
\arguments{
\item{mydata}{Matrix or data.frame with RNA-seq gene TPM or microarray
gene expression values for all samples to be deconvoluted,
with gene symbols as row names and sample IDs as column names.
Expression levels should be on non-log scale.}
}
\value{
TODO
}
\description{
mapGenes
}
\examples{
# TODO
}
