% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{check_signature}
\alias{check_signature}
\title{Check the signature matrix}
\usage{
check_signature(signature_matrix, mix_mat)
}
\arguments{
\item{signature_matrix}{A data.frame or a matrix object, containing the
signature matrix}

\item{mix_mat}{Mixture matrix, storing the information provided as \code{expression_data}
to the main function, \code{run_quantiseqr()}}
}
\value{
Invisible NULL
}
\description{
Checks requirements for the signature matrix, with respect to the expression
matrix data provided (the one on which the deconvolution algorithm needs to
be run)
}
\details{
Performs a number of checks to ensure the compatibility of the provided
signature matrix in \code{quantiseqr}, referring also to the content of the \code{mix_mat}
mixture matrix, to be deconvoluted
}
