% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{DClsei}
\alias{DClsei}
\title{Title}
\usage{
DClsei(b, A, G, H, scaling)
}
\arguments{
\item{b}{Numeric vector containing the right-hand side of the quadratic function to be minimised.}

\item{A}{Numeric matrix containing the coefficients of the quadratic function to be minimised.}

\item{G}{Numeric matrix containing the coefficients of the inequality constraints.}

\item{H}{Numeric vector containing the right-hand side of the inequality constraints.}

\item{scaling}{TODO}
}
\value{
TODO
}
\description{
Title
}
\details{
The \code{\link[limsolve:lsei]{limsolve::lsei()}} function is used as underlying framework. Please
refer to that function for more details.
}
\examples{
# TODO
}
