% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{DCrr}
\alias{DCrr}
\title{Title}
\usage{
DCrr(b, A, method, scaling)
}
\arguments{
\item{b}{TODO}

\item{A}{TODO}

\item{method}{TODO}

\item{scaling}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
