% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{extract_ti_from_se}
\alias{extract_ti_from_se}
\title{Extract tumor immune quantifications}
\usage{
extract_ti_from_se(se)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object, or any of its derivates, which
contains the quantifications extracted via \code{quantiseqr} in its \code{colData} slot.}
}
\value{
A data.frame, formatted as required by downstream functions
}
\description{
Extract tumor immune quantifications from a SummarizedExperiment object,
previously processed with \code{run_quantiseqr()}
}
\examples{
# TODO
}
