% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{fixMixture}
\alias{fixMixture}
\title{Title}
\usage{
fixMixture(mix.mat, arrays = FALSE)
}
\arguments{
\item{mix.mat}{TODO}

\item{arrays}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
