% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{quantiplot}
\alias{quantiplot}
\title{Plot the information on the tumor immune contexture}
\usage{
quantiplot(obj)
}
\arguments{
\item{obj}{An object, either
\itemize{
\item a \code{SummarizedExperiment} where the quantifications are stored
\item a simple data.frame object, as obtained by \code{run_quantiseqr()}
}}
}
\value{
A ggplot object
}
\description{
Plot the information on the tumor immune contexture, as extracted with
\code{run_quantiseqr()}
}
\examples{
# TODO
}
