% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{quanTIseq}
\alias{quanTIseq}
\title{Title}
\usage{
quanTIseq(currsig, currmix, scaling, method)
}
\arguments{
\item{currsig}{Signature matrix to be used for deconvolution (format: genes by cell types).}

\item{currmix}{Mixture matrix to be deconvoluted (format: genes by samples).}

\item{scaling}{Logical value. If set to FALSE, it disables the correction
of cell-type-specific mRNA content bias. Default: TRUE}

\item{method}{Character string, defining the deconvolution method to be used:
\code{lsei} for constrained least squares regression, \code{hampel}, \code{huber}, or \code{bisquare}
for robust regression with Huber, Hampel, or Tukey bisquare estimators,
respectively. Default: \code{lsei}.}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
