% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr_helpers.R
\name{quanTIseq}
\alias{quanTIseq}
\title{Title}
\usage{
quanTIseq(currsig, currmix, scaling, method)
}
\arguments{
\item{currsig}{TODO}

\item{currmix}{TODO}

\item{scaling}{TODO}

\item{method}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
