% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseqr.R
\name{run_quantiseq}
\alias{run_quantiseq}
\title{Use quanTIseq to deconvolute a gene expression matrix.}
\usage{
run_quantiseq(
  expression_data,
  arrays = FALSE,
  signame = "TIL10",
  tumor = FALSE,
  mRNAscale = TRUE,
  method = "lsei",
  column = "gene_symbol",
  rmgenes = "unassigned"
)
}
\arguments{
\item{expression_data}{table with the gene TPM or microarray expression values for all samples to be deconvoluted
(Gene symbols on the first column and sample IDs on the first row). Expression data should be on non-log scale}

\item{arrays}{Logical value. Set to TRUE if the expression data are from microarrays instead of RNA-seq. If TRUE, the "rmgenes" parameter is set to "none". Default: FALSE}

\item{signame}{name of the signature matrix. Currently only \code{TIL10} is available.
TODO: if passed as a file/matrix, we need to define some ways of handling this upstream of things}

\item{tumor}{Logical value. Set to TRUE if the expression data is from tumor samples. Default: FALSE}

\item{mRNAscale}{Logical value. Set to FALSE to disable the correction of cell-type-specific mRNA content bias. Default: TRUE}

\item{method}{Character string, defining the deconvolution method to be used:
"lsei" for constrained least squares regression, "hampel", "huber", or "bisquare"
for robust regression with Huber, Hampel, or Tukey bisquare estimators. Default: "lsei".}

\item{column}{Character, specifies which column contains the information of the gene symbol identifiers}

\item{rmgenes}{Specifies which genes have to be removed from the deconvolution analysis.
Can be a vector of gene symbols or a string among "none" (no genes are removed) and "default" (a list of genes with noisy expression RNA-seq data is removed as explained in quanTIseq paper). Default: "default" for RNA-seq data, "none" for microarrrays.}
}
\description{
Source code from https://github.com/FFinotello/quanTIseq - TODO: will need to update link?
}
\details{
F. Finotello, C. Mayer, C. Plattner, G. Laschober, D. Rieder,
H. Hackl, A. Krogsdam, Z. Loncova, W. Posch, D. Wilflingseder, S. Sopper, M. Jsselsteijn,
T. P. Brouwer, D. Johnsons, Y. Xu, Y. Wang, M. E. Sanders, M. V. Estrada, P. Ericsson-Gonzalez,
P. Charoentong, J. Balko, N. F. d. C. C. de Miranda, Z. Trajanoski.
"Molecular and pharmacological modulators of the tumor immune contexture revealed by deconvolution of RNA-seq data".
Genome Medicine 2019;11(1):34. doi: 10.1186/s13073-019-0638-6.
}
\examples{
# TODO

}
